requirements.txt
openai
python-dotenv
google-generativeai
pollinations
Pillow
speechrecognition
pygame
eel
pandas
requests
sounddevice
numpy